requirements.txt
requests
google-genai
pytz
requests_cache
PyYAML